# rotas-openapi.yaml
openapi: 3.0.4
info:
  title: APDIS - Rotas API
  description: |-
    Representação da arquitetura REST da plataforma APDIS, desenvolvida no âmbito da unidade curricular de Projeto de Software Web, Móvel e na Nuvem.
    Neste documento será possível observar as rotas existentes, os campos necessários e respostas esperadas.
  version: 1.0.0
servers:
  - url: https://projeto.matz.pt/api
...
paths:
...
  /documents/getAllRequests:
    get:
      tags:
        - documents
      summary: Obter todos os Pedidos
      description: Obtém a lista de todos os pedidos existentes.
      responses:
        '200':
          description: Lista de pedidos obtida com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    description:
                      type: string
                    requested:
                      type: string
                    accepted:
                      type: integer
                      nullable: true
                    documentName:
                      type: string
                      nullable: true
                    documentDOI:
                      type: string
                      nullable: true
                    documentData:
                      type: string
                      nullable: true
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
  /documents/getRequest:
    get:
      tags:
        - documents
      summary: Obter um Pedido pelo ID
      description: Retorna os detalhes de um pedido específico.
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
          description: ID do pedido a obter
      responses:
        '200':
          description: Pedido obtido com sucesso
          content:
            application/json:
              schema:
                type: object
        '400':
          description: ID de pedido inválido ou não autorizado
        '500':
          description: Erro interno
  /documents/createRequest:
    post:
      tags:
        - documents
      summary: Criar um novo Pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
              properties:
                description:
                  type: string
      responses:
        '200':
          description: Pedido criado com sucesso
        '400':
          description: Dados inválidos ou não autorizado
        '500':
          description: Erro interno
  /documents/acceptRequest:
    post:
      tags:
        - documents
      summary: Aceitar um Pedido
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - id
                - file
              properties:
                id:
                  type: integer
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Pedido aceite e documento enviado
        '400':
          description: Dados inválidos ou não autorizado
        '500':
          description: Erro interno
  /documents/recoverRequest:
    post:
      tags:
        - documents
      summary: Reverter um Pedido Aceite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: integer
      responses:
        '200':
          description: Pedido revertido com sucesso
        '400':
          description: Não autorizado ou ID inválido
        '500':
          description: Erro interno
  /documents/deleteRequest:
    delete:
      tags:
        - documents
      summary: Apagar um Pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: integer
      responses:
        '200':
          description: Pedido apagado com sucesso
        '400':
          description: Não autorizado ou ID inválido
        '500':
          description: Erro interno
  /documents/downloadRequest:
    get:
      tags:
        - documents
      summary: Download de Documento associado a um Pedido
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
          description: ID do pedido
      responses:
        '200':
          description: Download iniciado com sucesso
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Não autorizado ou ID inválido
        '404':
          description: Documento ou ficheiro não encontrado
        '500':
          description: Erro interno
...